[addons][player] move down the previous on addons used headers
This are no more used now by addons itself and only be used on
videoplayer itself.
