+++ xbmc/addons/AddonBindings.cmake
@@ -8,5 +8,4 @@ set(CORE_ADDON_BINDINGS_FILES
 
 set(CORE_ADDON_BINDINGS_DIRS
     ${CORE_SOURCE_DIR}/xbmc/addons/kodi-dev-kit/include/kodi/
-    ${CORE_SOURCE_DIR}/xbmc/cores/VideoPlayer/Interface/Addon
 )
