+++ xbmc/cores/VideoPlayer/CMakeLists.txt
@@ -36,9 +36,9 @@ set(HEADERS AudioSinkAE.h
             VideoPlayerVideo.h
             VideoReferenceClock.h
             Interface/StreamInfo.h
-            Interface/Addon/DemuxPacket.h
-            Interface/Addon/DemuxCrypto.h
-            Interface/Addon/InputStreamConstants.h
-            Interface/Addon/TimingConstants.h)
+            Interface/DemuxPacket.h
+            Interface/DemuxCrypto.h
+            Interface/InputStreamConstants.h
+            Interface/TimingConstants.h)
 
 core_add_library(VideoPlayer)
