+++ xbmc/cores/VideoPlayer/DVDCodecs/Audio/DVDAudioCodec.h
@@ -9,8 +9,8 @@
 #pragma once
 
 #include "cores/AudioEngine/Utils/AEAudioFormat.h"
+#include "cores/VideoPlayer/Interface/DemuxPacket.h"
 #include "cores/VideoPlayer/Process/ProcessInfo.h"
-#include "cores/VideoPlayer/Interface/Addon/DemuxPacket.h"
 
 #include <vector>
