+++ xbmc/cores/VideoPlayer/DVDDemuxers/DVDDemuxCC.cpp
@@ -10,7 +10,7 @@
 
 #include "DVDDemuxUtils.h"
 #include "cores/VideoPlayer/DVDCodecs/Overlay/contrib/cc_decoder708.h"
-#include "cores/VideoPlayer/Interface/Addon/TimingConstants.h"
+#include "cores/VideoPlayer/Interface/TimingConstants.h"
 
 #include <algorithm>
