+++ xbmc/cores/VideoPlayer/Interface/InputStreamConstants.h
