+++ xbmc/cores/VideoPlayer/VideoPlayerAudio.h
@@ -13,7 +13,7 @@
 #include "DVDMessageQueue.h"
 #include "DVDStreamInfo.h"
 #include "IVideoPlayer.h"
-#include "cores/VideoPlayer/Interface/Addon/TimingConstants.h"
+#include "cores/VideoPlayer/Interface/TimingConstants.h"
 #include "threads/Thread.h"
 #include "utils/BitstreamStats.h"
