+++ xbmc/cores/VideoPlayer/VideoPlayerRadioRDS.cpp
@@ -31,7 +31,7 @@
 #include "GUIUserMessages.h"
 #include "ServiceBroker.h"
 #include "cores/FFmpeg.h"
-#include "cores/VideoPlayer/Interface/Addon/TimingConstants.h"
+#include "cores/VideoPlayer/Interface/TimingConstants.h"
 #include "dialogs/GUIDialogKaiToast.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
