+++ xbmc/cores/VideoPlayer/VideoPlayerTeletext.cpp
@@ -9,8 +9,8 @@
 #include "VideoPlayerTeletext.h"
 
 #include "DVDStreamInfo.h"
-#include "cores/VideoPlayer/Interface/Addon/DemuxPacket.h"
-#include "cores/VideoPlayer/Interface/Addon/TimingConstants.h"
+#include "cores/VideoPlayer/Interface/DemuxPacket.h"
+#include "cores/VideoPlayer/Interface/TimingConstants.h"
 #include "threads/SingleLock.h"
 #include "utils/log.h"
